Refactor user , post and comment controllers: enhance user role validation, improve post and comment access control, and streamline user management logic

+++ src/controllers/post_controller.ts
@@ -1,10 +1,12 @@
 import postModel from "../models/post_model";
 import { Request, Response } from "express";
+import userModel from "../models/user_model";
 
 const addPost = async (req: Request, res: Response) => {
-  const postBody = req.body;
+  const userId = req.params.userId;
+  const { title, content } = req.body;
   try {
-    const post = await postModel.create(postBody);
+    const post = await postModel.create({ userID: userId, title, content });
     res.status(201).send(post);
   } catch (error) {
     res.status(400).send(error);
@@ -27,23 +29,39 @@ const getPost = async (req: Request, res: Response) => {
 };
 
 const getPostById = async (req: Request, res: Response) => {
-  const postId = req.params.id;
-
+  
+  const {postId} = req.params;
   try {
     const post = await postModel.findById(postId);
     if (post != null) {
+   
       res.send(post);
     } else {
       res.status(404).send("Post not found");
     }
+    
   } catch (error) {
     res.status(400).send(error);
   }
 };
 
 const deletePost = async (req: Request, res: Response) => {
-  const postId = req.params.id;
+  const {postId , userId} = req.params;
   try {
+    const post = await postModel.findById(postId);
+    const user = await userModel.findById(userId);
+    if(!user){
+      res.status(404).json({ message: 'User not found' });
+      return;
+    }
+    if (!post) {
+      res.status(404).json({ message: 'Post not found' });
+      return;
+    }
+    if(post.userID != userId && user.role != "admin"){
+      res.status(403).send("Access denied");
+      return;
+    }
     const rs = await postModel.findByIdAndDelete(postId);
     res.status(200).send(rs);
   } catch (error) {
@@ -52,18 +70,22 @@ const deletePost = async (req: Request, res: Response) => {
 };
 
 const updatePost = async (req: Request, res: Response) => { 
-  const id = req.params.id;
+  const {id, userId} = req.params;
   const body = req.body;
-
+  
   try {
     const post = await postModel.findById(id);
     if (!post) {
       res.status(404).json({ message: 'Post not found' });
+      return;
     }
-    else{
+    if(post.userID != userId){
+      res.status(403).send("Access denied");
+      return;
+    }
       const updatedPost = await postModel.findByIdAndUpdate(id, {userID: post.userID, title: body.title, content: body.content});
       res.status(200).send(updatedPost);
-    }
+    
   } catch (error) {
     res.status(500).send(error);
   }
